cmake: refactor message verbosity
This refactor boils down to changing the log level on overly detailed information
on `message(STATUS ...)` to `message(VERBOSE ...)` or `message(DEBUG ...)`,
this means by default they are not shown, and can be enabled with the CMake option
`–loglevel=VERBOSE/DEBUG` accordingly. Verbose messages contain detailed information
that may be relevant to the user, while debug messages are intended for the
developers to diagnose CMake configuring issues.